requirements.txt
flask==2.0.1
werkzeug==2.0.1
newspaper3k==0.2.8
nltk==3.6.3
beautifulsoup4==4.9.3
requests==2.26.0
numpy>=1.20.0
networkx>=2.6.3
python-dotenv>=0.19.0
transformers==4.18.0
torch==2.0.0
sentencepiece>=0.1.96
accelerate>=0.5.1
scikit-learn>=1.0.2
gunicorn==20.1.0
lxml[html_clean]>=4.9.2
tqdm>=4.65.0